+++ Source/Scene/ModelInstanceCollection.js
@@ -1,6 +1,8 @@
 define([
         '../Core/BoundingSphere',
+        '../Core/Cartesian2',
         '../Core/Cartesian3',
+        '../Core/Check',
         '../Core/clone',
         '../Core/Color',
         '../Core/ComponentDatatype',
@@ -28,7 +30,9 @@ define([
         './ShadowMode'
     ], function(
         BoundingSphere,
+        Cartesian2,
         Cartesian3,
+        Check,
         clone,
         Color,
         ComponentDatatype,
@@ -166,6 +170,10 @@ define([
 
         this.debugWireframe = defaultValue(options.debugWireframe, false);
         this._debugWireframe = false;
+
+        this._imageBasedLightingFactor = new Cartesian2(1.0, 1.0);
+        Cartesian2.clone(options.imageBasedLightingFactor, this._imageBasedLightingFactor);
+        this.lightColor = options.lightColor;
     }
 
     defineProperties(ModelInstanceCollection.prototype, {
@@ -192,6 +200,21 @@ define([
         readyPromise : {
             get : function() {
                 return this._readyPromise.promise;
+            }
+        },
+        imageBasedLightingFactor : {
+            get : function() {
+                return this._imageBasedLightingFactor;
+            },
+            set : function(value) {
+                //>>includeStart('debug', pragmas.debug);
+                Check.typeOf.object('imageBasedLightingFactor', value);
+                Check.typeOf.number.greaterThanOrEquals('imageBasedLightingFactor.x', value.x, 0.0);
+                Check.typeOf.number.lessThanOrEquals('imageBasedLightingFactor.x', value.x, 1.0);
+                Check.typeOf.number.greaterThanOrEquals('imageBasedLightingFactor.y', value.y, 0.0);
+                Check.typeOf.number.lessThanOrEquals('imageBasedLightingFactor.y', value.y, 1.0);
+                //>>includeEnd('debug');
+                Cartesian2.clone(value, this._imageBasedLightingFactor);
             }
         }
     });
@@ -576,7 +599,9 @@ define([
             uniformMapLoaded : undefined,
             pickIdLoaded : collection._pickIdLoaded,
             ignoreCommands : true,
-            opaquePass : collection._opaquePass
+            opaquePass : collection._opaquePass,
+            imageBasedLightingFactor : collection.imageBasedLightingFactor,
+            lightColor : collection.lightColor
         };
 
         if (!usesBatchTable) {
@@ -869,6 +894,9 @@ define([
 
         var instancingSupported = this._instancingSupported;
         var model = this._model;
+
+        model.imageBasedLightingFactor = this.imageBasedLightingFactor;
+        model.lightColor = this.lightColor;
 
         model.update(frameState);